--- a/app/api/boards/[id]/notes/route.ts
+++ b/app/api/boards/[id]/notes/route.ts
@@ -15,13 +15,25 @@ export async function GET(
 
     const board = await db.board.findUnique({
       where: { id: boardId },
-      include: { 
+      select: {
+        id: true,
+        isPublic: true,
+        organizationId: true,
         notes: {
           where: {
             deletedAt: null, // Only include non-deleted notes
             done: false
           },
-          include: {
+          select: {
+            id: true,
+            content: true,
+            color: true,
+            boardId: true,
+            createdBy: true,
+            createdAt: true,
+            updatedAt: true,
+            done: true,
+            checklistItems: true,
             user: {
               select: {
                 id: true,
@@ -29,7 +41,8 @@ export async function GET(
                 email: true
               }
             }
-          }
+          },
+          orderBy: { createdAt: 'desc' }
         }
       }
     })
@@ -48,7 +61,9 @@ export async function GET(
 
     const user = await db.user.findUnique({
       where: { id: session.user.id },
-      include: { organization: true }
+      select: { 
+        organizationId: true 
+      }
     })
 
     if (!user?.organizationId) {
@@ -83,7 +98,16 @@ export async function POST(
     // Verify user has access to this board (same organization)
     const user = await db.user.findUnique({
       where: { id: session.user.id },
-      include: { organization: true }
+      select: { 
+        organizationId: true,
+        organization: {
+          select: {
+            slackWebhookUrl: true
+          }
+        },
+        name: true,
+        email: true
+      }
     })
 
     if (!user?.organizationId) {