# What's changing?

* Refactored database queries across boards and notes endpoints by replacing `include` with explicit `select` field selection, which avoids unnecessary data fetching & makes more maintainable code

---

## 1) `GET /api/boards`

**Before**

```ts
const user = await db.user.findUnique({
  where: { id: session.user.id },
  include: { organization: true }  // Fetches entire organization object
})

// Later access: user.organization.id
```

**After**

```ts
const user = await db.user.findUnique({
  where: { id: session.user.id },
  select: { organizationId: true }  // Just the ID we need
})

// Direct access: user.organizationId
```

---

## 2) `GET /api/boards/[id]/notes`

**Before**

```ts
const board = await db.board.findUnique({
  where: { id: boardId },
  include: { 
    notes: {
      where: { deletedAt: null, done: false },
      include: {
        user: true  // All user fields
      }
    }
  }
})
```

**After**

```ts
const board = await db.board.findUnique({
  where: { id: boardId },
  select: {
    id: true,
    isPublic: true,
    organizationId: true,
    notes: {
      where: { deletedAt: null, done: false },
      select: {
        id: true,
        content: true,
        color: true,
        boardId: true,
        createdBy: true,
        createdAt: true,
        updatedAt: true,
        done: true,
        checklistItems: true,
        user: {
          select: { id: true, name: true, email: true }  // Only needed fields
        }
      },
      orderBy: { createdAt: 'desc' }
    }
  }
})
```

---

## 3) `POST /api/boards/[id]/notes`

**Before**

```ts
const user = await db.user.findUnique({
  where: { id: session.user.id },
  include: { organization: true }  // Fetches entire organization object
})
```

**After**

```ts
const user = await db.user.findUnique({
  where: { id: session.user.id },
  select: { 
    organizationId: true,
    organization: {
      select: {
        slackWebhookUrl: true  // Only the webhook URL needed
      }
    },
    name: true,
    email: true
  }
})
```

---

## 4) `GET /api/boards/archive/notes`

**Before**

```ts
const notes = await db.note.findMany({
  where: { /* ... */ },
  include: {
    user: true,   // All user fields
    board: true   // All board fields
  }
})
```

**After**

```ts
const notes = await db.note.findMany({
  where: { /* ... */ },
  select: {
    id: true,
    content: true,
    color: true,
    boardId: true,
    createdBy: true,
    createdAt: true,
    updatedAt: true,
    done: true,
    checklistItems: true,
    user: {
      select: { id: true, name: true, email: true }
    },
    board: {
      select: { id: true, name: true }
    }
  }
})
```

---

## Benchmarks

<details>
<summary>Benchmark Results</summary>

```bash
Performance Benchmark Comparison

GET /api/boards
  Unoptimized: 16.26ms | 2 queries | 2.4KB
  Optimized:   12.47ms | 2 queries | 2.4KB
  Improvement: 24.4% faster | 0.0% less data

GET /api/boards/[id]/notes
  Unoptimized: 12.88ms | 2 queries | 160.8KB
  Optimized:   11.38ms | 2 queries | 145.6KB
  Improvement: 11.3% faster | 9.5% less data

GET /api/boards/archive/notes
  Unoptimized: 19.68ms | 2 queries | 460.1KB
  Optimized:   17.93ms | 2 queries | 422.0KB
  Improvement: 8.8% faster | 8.3% less data

POST /api/boards/[id]/notes
  Unoptimized: 2.28ms | 5 queries | 0.4KB
  Optimized:   1.77ms | 5 queries | 0.4KB
  Improvement: 21.3% faster | 9.5% less data

Benchmark completed (100 cycles averaged)
```

Benchmark script can be found [here](https://gist.github.com/crisog/43375846012e7842446be5e321d88ccf).

*Note: These are local benchmarks with no network latency. In production environments with network overhead, the relative improvements may be less pronounced but data transfer savings will be more significant.*

</details>

## Impact

The primary benefits are:
- **Reduced data transfer**: Less bandwidth usage and faster response parsing, especially valuable over slower network connections
- **Database efficiency**: Eliminates unnecessary field fetching, reducing database load 
- **Maintainability**: Explicit field selection makes it clear what data each endpoint actually needs

While the local performance gains are marginal, the cumulative effect of reduced data transfer and database overhead becomes more significant in real-world network conditions.

## Frontend Compatibility

All API endpoint changes maintain **100% backwards compatibility** with the frontend:

| Endpoint | Response Format | Field Compatibility |
|----------|----------------|-------------------|
| `GET /api/boards` | `{ boards }` (unchanged) | All dashboard fields preserved: `id`, `name`, `description`, `isPublic`, `createdBy`, `_count.notes` |
| `GET /api/boards/[id]/notes` | `{ notes }` (unchanged) | All Note interface fields preserved: `id`, `content`, `color`, `done`, `createdAt`, `updatedAt`, `checklistItems`, `boardId`, `user.{id,name,email}` |
| `POST /api/boards/[id]/notes` | `{ note }` (unchanged) | Note creation unchanged - uses `include` with proper field selection |
| `GET /api/boards/archive/notes` | `{ notes }` (unchanged) | All fields preserved including `board.{id,name}` for archive display |
| `POST /api/boards` | `{ board }` (unchanged) | All board creation fields preserved: `id`, `name`, `description`, `isPublic`, `createdBy`, `_count.notes` |

The optimizations only change internal database query structure while maintaining identical response formats for the frontend.

E2E tests are passing.

<img width="414" height="243" alt="image" src="https://github.com/user-attachments/assets/ed6759cf-e1bc-47d0-bf69-0088cc769bf7" />